Refactor teaser popup component: replace FameOnMePopup with TeaserPopup and remove unused file
