+++ runningdinner-client/webapp/src/common/teaserpopup/index.ts
@@ -1,0 +1,1 @@
+export * from './TeaserPopup';
