+++ runningdinner-client/webapp/src/self/SelfAdminApp.tsx
@@ -8,7 +8,7 @@ import {
 import { SelfAdminRoute } from './SelfAdminRoute';
 import {selfAdminStore} from "@runningdinner/shared";
 import { ProgressBar } from '../common/ProgressBar';
-import { FameOnMePopup } from '../common/teaserpopup/FameOnMePopup';
+import { TeaserPopup } from '../common/teaserpopup';
 
 export default function SelfAdminApp() {
 
@@ -29,7 +29,7 @@ function SelfAdminAppPage() {
       <ProgressBar fetchError={fetchError} showLoadingProgress={showLoadingProgress} />
       <Container maxWidth={false}>
         <SelfAdminRoute />
-        <FameOnMePopup />
+        <TeaserPopup />
       </Container>
     </div>
   );
